requirements.txt
certifi==2024.7.4
coverage==7.6.0
feedparser==6.0.11
python-dateutil==2.9.0.post0
pytz==2024.1
sgmllib3k==1.0.0
six==1.16.0
